[CHE-280] Gather all error enums in a single file. Comment out unused errors. Rename .error(_ error: Error) to .underlyingError(_ error: Error)
