+++ PrimerSDK/Services/Error Handler/ErrorHandler.swift
@@ -1,0 +1,8 @@
+//
+//  ErrorHandler.swift
+//  PrimerSDK
+//
+//  Created by Evangelos Pittas on 16/3/21.
+//
+
+import Foundation
