+++ PrimerSDK/Services/Network/NetworkServiceError.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-enum NetworkServiceError: Error {
-    case invalidURL
-    case noData
-    case parsing(_ error: Error, _ data: Data)
-    case error(_ error: Error)
-}
-
 extension NetworkServiceError: LocalizedError {
     var errorDescription: String? {
         switch self {
@@ -24,7 +17,7 @@ extension NetworkServiceError: LocalizedError {
         case .parsing(let error, let data):
             let response = String(data: data, encoding: .utf8) ?? ""
             return "Parsing error: \(error.localizedDescription)\n\nResponse:\n\(response)"
-        case .error(let error):
+        case .underlyingError(let error):
             return error.localizedDescription
         }
     }
